+++ phils_model/src/json.h
@@ -4,16 +4,16 @@
 
 void save_weights_as_json(
     char *fname,
-    double ***weights_result,
-    double *layer_sizes,
+    float ***weights_result,
+    float *layer_sizes,
     int layer_sizes_rows,
     int layer_sizes_cols
 );
 
 void save_biases_as_json(
     char *fname,
-    double **biases,
-    double *layer_sizes,
+    float **biases,
+    float *layer_sizes,
     int layer_sizes_rows,
     int layer_sizes_cols
 );
